requirements.txt
pandas==0.24.0
numpy==1.16.1
scikit-learn==0.20.0
spacy==2.0.18
flask==1.0.2
tensorflow-gpu==1.13.1
